Extract Analyzable and Representable concerns
