+++ activestorage/lib/active_storage/errors.rb
@@ -1,0 +1,5 @@
+module ActiveStorage
+  class InvariableError < StandardError; end
+  class UnpreviewableError < StandardError; end
+  class UnrepresentableError < StandardError; end
+end
